+++ changelog.md
@@ -8,6 +8,8 @@ Codion Change Log
 ### is.codion.framework.db
 - EntityConnection.rollbackTransaction() and commitTransaction() no longer ignore exceptions, now throw DatabaseException.
 - EntityConnection.transaction() added along with Transactional and TransactionalResult.
+### is.codion.framework.db.local
+- DefaultLocalEntityConnection.valueMissingOrModified() bug fixed, could not discern a missing value from a null original value, since the incorrent entity was being used when checking, test case improved.
 ### is.codion.framework.model
 - EntityConditions.AdditionalCondition moved to EntityQueryModel, additionalWhere() and additionalHaving() renamed where() and having() respectively.
 ### is.codion.swing.framework.ui